# Upgrade to latest mimir-prometheus