… go.sum
… vendor/github.com/prometheus/prometheus/discovery/metrics.go
… vendor/github.com/prometheus/prometheus/discovery/metrics_k8s_client.go
… vendor/github.com/prometheus/prometheus/model/labels/regexp.go
… vendor/github.com/prometheus/prometheus/promql/query_logger.go
… vendor/github.com/prometheus/prometheus/rules/alerting.go
… vendor/github.com/prometheus/prometheus/scrape/manager.go
… vendor/github.com/prometheus/prometheus/scrape/scrape.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheus/normalize_label.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheus/normalize_name.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheus/unit_to_ucum.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite/helper.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite/histograms.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite/metrics_to_prw.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite/number_data_points.go
… vendor/github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite/otlp_to_openmetrics_metadata.go
… vendor/github.com/prometheus/prometheus/storage/remote/write_handler.go
… vendor/modules.txt