#### What this PR does

Upgrade to latest mimir-prometheus: [de14b0ea393d](https://github.com/grafana/mimir-prometheus/tree/de14b0ea393dbe5c3c93965e754120ea6d476909).

#### Significant Prometheus changes

* [OTLP: Use PrometheusConverter directly](https://github.com/prometheus/prometheus/pull/14006)
* [otlp: Prometheus to own its own copy of the otlptranslator package](https://github.com/prometheus/prometheus/pull/13991)
* [Remove OpEmptyMatch from regex concatenations](https://github.com/grafana/mimir-prometheus/pull/615)
* [OTLP: Don't generate target_info unless at least one identifying label is defined.](https://github.com/prometheus/prometheus/pull/13991)
* [OTLP: Don't generate target_info unless there are metrics.](https://github.com/prometheus/prometheus/pull/13991)

The main changes are OTLP related, i.e. that Prometheus now has taken ownership of its `otlptranslator` package (forked it from opentelemetry-collector-contrib), and that the OTel->Prometheus remote write format translation is now optimized through hash based metric identifiers (instead of string based ones). Prometheus [benchmarks](https://github.com/prometheus/prometheus/pull/14006) show huge performance improvements from the latter.

#### Benchmarks

The `OTLPHandler` benchmarks don't show any enormous improvements, but it's possible that this is due to having to convert from Prometheus remote write format to Mimir's. It's also possible that our benchmark suite doesn't quite cover all of the improved cases versus the benchmark suite in Prometheus.

<details>
<summary>Benchmark stats vs main branch</summary>

```console
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/distributor
                        │ main-otlphandler-6.txt │     upgrade-otlphandler-6.txt     │
                        │         sec/op         │   sec/op     vs base              │
OTLPHandler/protobuf-12              262.2µ ± 1%   252.8µ ± 0%  -3.60% (p=0.002 n=6)
OTLPHandler/JSON-12                  384.2µ ± 0%   376.2µ ± 0%  -2.09% (p=0.002 n=6)
geomean                              317.4µ        308.4µ       -2.85%

                        │ main-otlphandler-6.txt │     upgrade-otlphandler-6.txt      │
                        │          B/op          │     B/op      vs base              │
OTLPHandler/protobuf-12             393.8Ki ± 0%   370.1Ki ± 0%  -6.03% (p=0.002 n=6)
OTLPHandler/JSON-12                 440.7Ki ± 0%   416.9Ki ± 0%  -5.41% (p=0.002 n=6)
geomean                             416.6Ki        392.8Ki       -5.72%

                        │ main-otlphandler-6.txt │     upgrade-otlphandler-6.txt      │
                        │       allocs/op        │  allocs/op   vs base               │
OTLPHandler/protobuf-12              6.133k ± 0%   5.127k ± 0%  -16.40% (p=0.002 n=6)
OTLPHandler/JSON-12                  9.167k ± 0%   8.160k ± 0%  -10.99% (p=0.002 n=6)
geomean                              7.498k        6.468k       -13.74%
```

</details>

#### Which issue(s) this PR fixes or relates to

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.